+++ examples/csvimport_orm.py
@@ -2,16 +2,14 @@
 A simple example that imports a tab-delimited CSV file (people.csv) into an SQLite database
 (people.db).
 
-This version of the example uses the ORM instead of core queries.
+This version of the example uses SQLAlchemy ORM.
 """
 
-import asyncio
 import csv
 import logging
 from pathlib import Path
 
-from asphalt.core import ContainerComponent, Context, run_application
-from asyncio_extras.threads import threadpool
+from asphalt.core import CLIApplicationComponent, Context, run_application
 from sqlalchemy.ext.declarative.api import declarative_base
 from sqlalchemy.sql.schema import Column
 from sqlalchemy.sql.sqltypes import Integer, Unicode
@@ -30,33 +28,37 @@ class Person(Base):
     email = Column(Unicode, nullable=False)
 
 
-class ApplicationrComponent(ContainerComponent):
-    async def start(self, ctx: Context):
-        csv_path = Path(__file__).with_name('people.csv')
-        db_path = csv_path.with_name('people.db')
+class CSVImporterComponent(CLIApplicationComponent):
+    def __init__(self):
+        super().__init__()
+        self.csv_path = Path(__file__).with_name('people.csv')
 
-        # Remove any existing db file
+    async def start(self, ctx: Context):
+        # Remove the db file if it exists
+        db_path = self.csv_path.with_name('people.db')
         if db_path.exists():
             db_path.unlink()
 
-        self.add_component('sqlalchemy', url='sqlite:///{}'.format(db_path),
-                           metadata=Base.metadata)
+        self.add_component('sqlalchemy', url='sqlite:///{}'.format(db_path))
         await super().start(ctx)
 
-        # Create the table
-        Base.metadata.create_all()
+        # Create the table in a subcontext – never use connections or sessions from a long lived
+        # context!
+        async with Context(ctx):
+            Base.metadata.create_all(ctx.sql)
 
-        async with threadpool():
+    async def run(self, ctx: Context):
+        async with ctx.threadpool():
             num_rows = 0
-            with csv_path.open() as csvfile:
+            with self.csv_path.open() as csvfile:
                 reader = csv.reader(csvfile, delimiter='|')
-
-                # Import each row into the session as a new Person instance
                 for name, city, phone, email in reader:
                     num_rows += 1
                     ctx.dbsession.add(Person(name=name, city=city, phone=phone, email=email))
 
-        logger.info('Imported %d rows of data', num_rows)
-        asyncio.get_event_loop().stop()
+            # Emit pending INSERTs (though this would happen when the context ends anyway)
+            ctx.dbsession.flush()
 
-run_application(ApplicationrComponent(), logging=logging.DEBUG)
+        logger.info('Imported %d rows of data', num_rows)
+
+run_application(CSVImporterComponent(), logging=logging.DEBUG)